fix: create lc_704_binary_search.test.js with test cases
